Move utility function
[review feedback]
